01_Portfolios/app-dys/01_transition/requirements.txt
pygame
python-dotenv
supabase
gTTS
gotrue
httpx
pydantic
pydantic-settings
postgrest
storage3
realtime
typing-extensions
